Add files via upload
SOP (Search Optimization Pro) sells online digital advertising to local business via its sales associates. When a sales associate approaches a local business, he or she tries to convince the local business to let SOP use its experience to manage the business's online advertisements on Google, Facebook, etc. If the sales associate is successful, the business signs a contract with SOP, and they agree on the terms. 
The contract includes following details: 
•Monthly advertising budget (b), this is the maximum amount SOP will spend on behalf of the business each month. 
•Digital Advertising Setup Fee, Annual (s): $3000-5000, negotiable and varies according to each contract's terms and the business's complexity.
•Monthly Management Fee (f): $2500 + 15% of Monthly Ad Spend f = 2500 + 0.15 * b
•Based on the terms of the contract we can then estimate the total annual revenue (TAR) earned from a new business as:  TAR = s + 12 * f
•Once a contract is signed and underway, SOP pays annual commission (c) to its sales associate who delivered the contract. The commission amount is set as flat 5% of the total annual revenue of the contract. c = TAR * 0.05

You are asked to write a program that SOP can use to calculate f, TAR and c given b and s for each contracted sale. Below are the specifics.